Simplification and some project structure
